♻️ Refactor: 공통 컴포넌트 product card로 변경
