+++ src/widgets/profile/components/product-card-section/products/CreatedProducts.tsx
@@ -1,12 +1,29 @@
 import { Suspense } from 'react'
-import ProductList from './ProductsList'
+import { useIntersect } from '@shared/hooks'
 import { useFetchCreatedProducts } from '@shared/hooks/query'
+import { ProductCardList } from '@shared/ui'
 import type { UserId } from '@shared/types'
 
 export default function CreatedProductsList({ userId }: UserId) {
+  const {
+    data: products,
+    isFetching,
+    hasNextPage,
+    fetchNextPage,
+    error,
+  } = useFetchCreatedProducts({ userId })
+
+  if (error && !isFetching) throw error
+
+  const onIntersect = () => {
+    if (hasNextPage && !isFetching) fetchNextPage()
+  }
+
+  const ref = useIntersect<HTMLDivElement>(onIntersect)
+
   return (
     <Suspense fallback={<div></div>}>
-      <ProductList userId={userId} useFetchProducts={useFetchCreatedProducts} />
+      <ProductCardList data={products} ref={ref} />
     </Suspense>
   )
 }
